admin base functionality added

+++ src/main/java/com/controller/commands/admin/author/AdminRemoveAuthorCommand.java
@@ -3,8 +3,10 @@ package com.controller.commands.admin.author;
 import com.controller.commands.Command;
 import com.controller.commands.CommandWrapper;
 import com.controller.responce.Dispatcher;
+import com.controller.responce.EmptyDispatcher;
 import com.controller.responce.RedirectDispatcher;
 import com.service.AuthorService;
+import com.service.BookService;
 import com.service.impl.ServiceFactory;
 
 import javax.servlet.http.HttpServletRequest;
@@ -19,9 +21,22 @@ import static com.controller.constants.UrlsConst.ADMIN_AUTHORS;
 public class AdminRemoveAuthorCommand extends CommandWrapper implements Command {
     @Override
     protected Dispatcher processExecute(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        Integer authorId=Integer.valueOf(request.getParameter("authorId"));
+        Integer authorId=Integer.valueOf(request.getParameter("author_id"));
+        BookService bookService= ServiceFactory.getInstance().getBookService();
+
+        int relatedBooksCount=bookService.getBooksCountByParams(null,authorId,null,null,
+                null);
+
+        //check if author has some books
+        if(relatedBooksCount>0){
+            response.setStatus(HttpServletResponse.SC_FORBIDDEN);
+            return new EmptyDispatcher();
+        }
+
         AuthorService authorService= ServiceFactory.getInstance().getAuthorService();
         authorService.deleteById(authorId);
-        return new RedirectDispatcher(ADMIN_AUTHORS);
+
+        response.setStatus(HttpServletResponse.SC_OK);
+        return new EmptyDispatcher();
     }
 }
